Refactor rename models
