+++ .env.template
@@ -1,2 +1,3 @@
-GITHUB_TOKEN=
-PORT=
+GITHUB_TOKEN=<add your github api token here>
+PORT=<optional custom port to run the server on>
+
