+++ src/models/Assets.ts
