+++ src/models/Networks.ts
