+++ src/models/TokenRegistryServiceCache.ts
@@ -15,3 +15,4 @@ export type TokenRegistryServiceCache = {
     shimmer: CacheData,
     iota: CacheData
 }
+
